+++ app/helper/scaling.js
@@ -1,0 +1,18 @@
+import { Dimensions } from 'react-native';
+
+const { width, height } = Dimensions.get('window');
+
+// Guideline sizes are based on standard ~5" screen mobile device
+const guidelineBaseWidth = 360;
+const guidelineBaseHeight = 640;
+
+const scale = size => {
+  return (width / guidelineBaseWidth) * size;
+};
+// should renamed as horizontalScale
+// since scale has been used in several place will
+const horizontalScale = size => (width / guidelineBaseWidth) * size;
+const verticalScale = size => (height / guidelineBaseHeight) * size;
+const moderateScale = (size, factor = 0.5) => size + (scale(size) - size) * factor;
+
+export { scale, horizontalScale, verticalScale, moderateScale };
